Fix ufo_privilege_escalation
